Autogenerate library
Created a `repository_dispatch` event workflow that will be used to
generate, publish, and release the `mx-platform-ruby` library.

The workflow will accept a request from the `openapi` repo where the
body of the request includes the version bump (major, minor, patch).
This will allow us to generate and publish a new version of library when
changes are made to the openapi spec.