Simplify generate sdk workflow
Our `publish` and `release` workflows run when we `push` to `master` so we can remove the `publish` and `release` jobs from our repository dispatch `generate_sdk` workflow.

This should also fix any issues we are seeing where we are trying to publish or release using the wrong version. The old version seems to have been cached during the old flow which caused the workflow to fail.